requirements.txt
python-telegram-bot==20.7
psutil==5.9.6
aiofiles==23.2.1
aiohttp==3.9.1
matplotlib==3.8.2
pyjwt==2.8.0
asyncio==3.4.3
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
WMI==1.5.1 